fix file layout of visualization test
