+++ tests/test_metrics_constraintplot.py
@@ -24,46 +24,6 @@ def pre_post_process():
         shutil.rmtree(norm_path)
 
 
-def solve():
-    d = jm.Placeholder("d", dim=2)
-    x = jm.Binary("x", shape=d.shape)
-    i = jm.Element("i", d.shape[0])
-    j = jm.Element("j", d.shape[1])
-    problem = jm.Problem("simple_problem")
-    problem += jm.Sum([i, j], d[i, j] * x[i, j])
-    problem += jm.Constraint("onehot1", x[i, :] == 1, forall=i)
-    problem += jm.Constraint("onehot2", x[:, j] == 1, forall=j)
-    jm_sampleset_dict = {
-        "record": {
-            "solution": {
-                "x": [
-                    (([0, 1], [0, 1]), [1, 1], (2, 2)),
-                    (([0, 1], [1, 0]), [1, 1], (2, 2)),
-                    (([], []), [], (2, 2)),
-                    (([0, 1], [0, 0]), [1, 1], (2, 2)),
-                ]
-            },
-            "num_occurrences": [4, 3, 2, 1],
-        },
-        "evaluation": {
-            "energy": [3.0, 24.0, 0.0, 20.0],
-            "objective": [3.0, 24.0, 0.0, 17.0],
-            "constraint_violations": {
-                "onehot1": [0.0, 0.0, 2.0, 0.0],
-                "onehot2": [0.0, 0.0, 2.0, 2.0],
-            },
-            "penalty": {},
-        },
-        "measuring_time": {"solve": None, "system": None, "total": None},
-    }
-    jm_sampleset = jm.SampleSet.from_serializable(jm_sampleset_dict)
-    solving_time = jm.SolvingTime(
-        **{"preprocess": 1.0, "solve": 1.0, "postprocess": 1.0}
-    )
-    jm_sampleset.measuring_time.solve = solving_time
-    return jm_sampleset
-
-
 def simple_func_for_boxplot(x: pd.Series) -> dict:
     return {"data1": np.array([1, 2, 3, 4]), "data2": np.array([1, 2, 3, 4])}
 
@@ -79,10 +39,13 @@ def test_metrics_plot_get_violations_dict():
     assert (violations_dict["onehot_violations"] == np.array([0, 2])).all()
 
 
-def test_metrics_plot_boxplot_violations_return_value():
+def test_metrics_plot_boxplot_violations_return_value(jm_sampleset: jm.SampleSet):
+    def func() -> jm.SampleSet:
+        return jm_sampleset
+
     num_multipliers = 2
     params = {"multipliers": [{} for _ in range(num_multipliers)]}
-    solver_list = [solve]
+    solver_list = [func]
 
     expect_num_fig = len(params["multipliers"]) * len(solver_list)
 
@@ -98,12 +61,17 @@ def test_metrics_plot_boxplot_violations_return_value():
     assert type(fig_ax_tuple[0][1]) == axes.Subplot
 
 
-def test_metrics_plot_boxplot_violations_call_maplotlib_boxplot(mocker):
+def test_metrics_plot_boxplot_violations_call_maplotlib_boxplot(
+    mocker, jm_sampleset: jm.SampleSet
+):
+    def func() -> jm.SampleSet:
+        return jm_sampleset
+
     m = mocker.patch("matplotlib.axes.Subplot.boxplot")
 
     num_multipliers = 2
     params = {"multipliers": [{} for _ in range(num_multipliers)]}
-    solver_list = [solve]
+    solver_list = [func]
 
     expect_num_call = len(params["multipliers"]) * len(solver_list)
 
@@ -118,12 +86,15 @@ def test_metrics_plot_boxplot_violations_call_maplotlib_boxplot(mocker):
     assert m.call_count == expect_num_call
 
 
-def test_metrics_plot_boxplot_violations_arg_figsize():
+def test_metrics_plot_boxplot_violations_arg_figsize(jm_sampleset: jm.SampleSet):
+    def func() -> jm.SampleSet:
+        return jm_sampleset
+
     figwidth, figheight = 8, 4
 
     bench = jb.Benchmark(
         params={},
-        solver=[solve],
+        solver=[func],
     )
     result = bench()
     mplot = ConstraintPlot(result)
@@ -136,14 +107,17 @@ def test_metrics_plot_boxplot_violations_arg_figsize():
     assert fig.get_figheight() == 4
 
 
-def test_metrics_plot_boxplot_violations_arg_title(mocker):
+def test_metrics_plot_boxplot_violations_arg_title(mocker, jm_sampleset):
+    def func() -> jm.SampleSet:
+        return jm_sampleset
+
     m = mocker.patch("matplotlib.figure.Figure.suptitle")
 
     title = ["title1", "title2"]
 
     num_multipliers = 2
     params = {"multipliers": [{} for _ in range(num_multipliers)]}
-    solver_list = [solve]
+    solver_list = [func]
 
     bench = jb.Benchmark(
         params=params,
@@ -162,7 +136,12 @@ def test_metrics_plot_boxplot_violations_arg_title(mocker):
     )
 
 
-def test_metrics_plot_boxplot_violations_arg_title_fontsize(mocker):
+def test_metrics_plot_boxplot_violations_arg_title_fontsize(
+    mocker, jm_sampleset: jm.SampleSet
+):
+    def func() -> jm.SampleSet:
+        return jm_sampleset
+
     m = mocker.patch("matplotlib.figure.Figure.suptitle")
 
     title = ["title1", "title2"]
@@ -170,7 +149,7 @@ def test_metrics_plot_boxplot_violations_arg_title_fontsize(mocker):
 
     num_multipliers = 2
     params = {"multipliers": [{} for _ in range(num_multipliers)]}
-    solver_list = [solve]
+    solver_list = [func]
 
     bench = jb.Benchmark(
         params=params,
@@ -190,10 +169,13 @@ def test_metrics_plot_boxplot_violations_arg_title_fontsize(mocker):
     )
 
 
-def test_metrics_plot_boxplot_violations_arg_xticklabels():
+def test_metrics_plot_boxplot_violations_arg_xticklabels(jm_sampleset: jm.SampleSet):
+    def func() -> jm.SampleSet:
+        return jm_sampleset
+
     bench = jb.Benchmark(
         params={},
-        solver=[solve],
+        solver=[func],
     )
     result = bench()
     mplot = ConstraintPlot(result)
@@ -205,13 +187,18 @@ def test_metrics_plot_boxplot_violations_arg_xticklabels():
     assert xticklabels[1].get_text() == "onehot2_violations"
 
 
-def test_metrics_plot_boxplot_violations_arg_xticklabels_size(mocker):
+def test_metrics_plot_boxplot_violations_arg_xticklabels_size(
+    mocker, jm_sampleset: jm.SampleSet
+):
+    def func() -> jm.SampleSet:
+        return jm_sampleset
+
     m = mocker.patch("matplotlib.axes.Subplot.set_xticklabels")
     constraint_name_fontsize = 10
 
     bench = jb.Benchmark(
         params={},
-        solver=[solve],
+        solver=[func],
     )
     result = bench()
     mplot = ConstraintPlot(result)
@@ -222,13 +209,18 @@ def test_metrics_plot_boxplot_violations_arg_xticklabels_size(mocker):
     assert kwargs["size"] == 10
 
 
-def test_metrics_plot_boxplot_violations_arg_xticklabels_rotation(mocker):
+def test_metrics_plot_boxplot_violations_arg_xticklabels_rotation(
+    mocker, jm_sampleset: jm.SampleSet
+):
+    def func() -> jm.SampleSet:
+        return jm_sampleset
+
     m = mocker.patch("matplotlib.axes.Subplot.set_xticklabels")
     constraint_name_fontrotation = 10
 
     bench = jb.Benchmark(
         params={},
-        solver=[solve],
+        solver=[func],
     )
     result = bench()
     mplot = ConstraintPlot(result)
@@ -239,10 +231,13 @@ def test_metrics_plot_boxplot_violations_arg_xticklabels_rotation(mocker):
     assert kwargs["rotation"] == 10
 
 
-def test_metrics_plot_boxplot_violations_arg_ylabel_default():
+def test_metrics_plot_boxplot_violations_arg_ylabel_default(jm_sampleset: jm.SampleSet):
+    def func() -> jm.SampleSet:
+        return jm_sampleset
+
     bench = jb.Benchmark(
         params={},
-        solver=[solve],
+        solver=[func],
     )
     result = bench()
     mplot = ConstraintPlot(result)
@@ -252,12 +247,15 @@ def test_metrics_plot_boxplot_violations_arg_ylabel_default():
     assert ax.get_ylabel() == "constraint violations"
 
 
-def test_metrics_plot_boxplot_violations_arg_ylabel():
+def test_metrics_plot_boxplot_violations_arg_ylabel(jm_sampleset: jm.SampleSet):
+    def func() -> jm.SampleSet:
+        return jm_sampleset
+
     ylabel = "ylabel"
 
     bench = jb.Benchmark(
         params={},
-        solver=[solve],
+        solver=[func],
     )
     result = bench()
     mplot = ConstraintPlot(result)
@@ -269,7 +267,12 @@ def test_metrics_plot_boxplot_violations_arg_ylabel():
     assert ax.get_ylabel() == "ylabel"
 
 
-def test_metrics_plot_boxplot_violations_arg_ylabel_size(mocker):
+def test_metrics_plot_boxplot_violations_arg_ylabel_size(
+    mocker, jm_sampleset: jm.SampleSet
+):
+    def func() -> jm.SampleSet:
+        return jm_sampleset
+
     m = mocker.patch("matplotlib.axes.Subplot.set_ylabel")
 
     ylabel = "ylabel"
@@ -277,7 +280,7 @@ def test_metrics_plot_boxplot_violations_arg_ylabel_size(mocker):
 
     bench = jb.Benchmark(
         params={},
-        solver=[solve],
+        solver=[func],
     )
     result = bench()
     mplot = ConstraintPlot(result)
@@ -286,13 +289,18 @@ def test_metrics_plot_boxplot_violations_arg_ylabel_size(mocker):
     m.assert_called_once_with("ylabel", size=ylabel_size)
 
 
-def test_metrics_plot_boxplot_violations_arg_yticks_default(mocker):
+def test_metrics_plot_boxplot_violations_arg_yticks_default(
+    mocker, jm_sampleset: jm.SampleSet
+):
+    def func() -> jm.SampleSet:
+        return jm_sampleset
+
     m = mocker.patch("matplotlib.ticker.MaxNLocator.__init__", return_value=None)
     mocker.patch("matplotlib.axes.Subplot.axhline")
 
     bench = jb.Benchmark(
         params={},
-        solver=[solve],
+        solver=[func],
     )
     result = bench()
     mplot = ConstraintPlot(result)
@@ -301,12 +309,15 @@ def test_metrics_plot_boxplot_violations_arg_yticks_default(mocker):
     m.assert_called_with(integer=True)
 
 
-def test_metrics_plot_boxplot_violations_arg_yticks():
+def test_metrics_plot_boxplot_violations_arg_yticks(jm_sampleset: jm.SampleSet):
+    def func() -> jm.SampleSet:
+        return jm_sampleset
+
     yticks = [0, 10]
 
     bench = jb.Benchmark(
         params={},
-        solver=[solve],
+        solver=[func],
     )
     result = bench()
     mplot = ConstraintPlot(result)
@@ -317,12 +328,17 @@ def test_metrics_plot_boxplot_violations_arg_yticks():
     assert (ax.get_yticks() == [0, 10]).all()
 
 
-def test_metrics_plot_boxplot_violations_arg_kwargs_default(mocker):
+def test_metrics_plot_boxplot_violations_arg_kwargs_default(
+    mocker, jm_sampleset: jm.SampleSet
+):
+    def func() -> jm.SampleSet:
+        return jm_sampleset
+
     m = mocker.patch("matplotlib.axes.Subplot.boxplot")
 
     bench = jb.Benchmark(
         params={},
-        solver=[solve],
+        solver=[func],
     )
     result = bench()
     mplot = ConstraintPlot(result)
@@ -333,12 +349,15 @@ def test_metrics_plot_boxplot_violations_arg_kwargs_default(mocker):
     assert kwargs["whis"] == [0, 100]
 
 
-def test_metrics_plot_boxplot_violations_arg_kwargs(mocker):
+def test_metrics_plot_boxplot_violations_arg_kwargs(mocker, jm_sampleset: jm.SampleSet):
+    def func() -> jm.SampleSet:
+        return jm_sampleset
+
     m = mocker.patch("matplotlib.axes.Subplot.boxplot")
 
     bench = jb.Benchmark(
         params={},
-        solver=[solve],
+        solver=[func],
     )
     result = bench()
     mplot = ConstraintPlot(result)
@@ -352,14 +371,17 @@ def test_metrics_plot_boxplot_violations_arg_kwargs(mocker):
     assert kwargs["meanline"] is True
 
 
-def test_metrics_plot_boxplot_violations_hline_indicate_no_violation(mocker):
+def test_metrics_plot_boxplot_violations_hline_indicate_no_violation(
+    mocker, jm_sampleset: jm.SampleSet
+):
+    def func() -> jm.SampleSet:
+        return jm_sampleset
+
     m = mocker.patch("matplotlib.axes.Subplot.axhline")
 
     num_multipliers = 2
     params = {"multipliers": [{} for _ in range(num_multipliers)]}
-    solver_list = [solve]
-
-    expect_num_call = len(params["multipliers"]) * len(solver_list)
+    solver_list = [func]
 
     bench = jb.Benchmark(
         params=params,
